Moved the test command to its own separate file : from queries.js to testqueries.js
