+++ queries.js
@@ -49,17 +49,7 @@ async function createUser(name) {
         console.log(err);
     }
 }
-/**
- * Clears the database. Used for tests only
- */
-async function clearUsers() {
-    const text = "TRUNCATE TABLE users CASCADE";
-    try {
-        await db.query(text);
-    } catch (err) {
-        console.log(err);
-    }
-}
+
 
 /**
  *
@@ -122,7 +112,6 @@ async function markAlbumAsPassed(albumName) {
 
 module.exports = {
     createUser,
-    clearUsers,
     incrementUserAlbumCount,
     createAlbumForUser,
     markAlbumAsPassed,
